feat: 관리자페이지 재난제보 평가/신고 포함해서 한번에 조회 & 신고 수 long -> int

+++ core-api/src/main/java/Hongik_SafeMap_Server/domain/disaster_report/dto/response/DisasterReportStatisticsResponse.java
@@ -21,12 +21,12 @@ public class DisasterReportStatisticsResponse {
     private int notHelpfulCount;
     
     @Schema(description = "총 신고 수", example = "2")
-    private long accusationCount;
+    private int accusationCount;
 
     public static DisasterReportStatisticsResponse of(
             Long disasterReportId,
             DisasterReportEvaluation evaluation,
-            long accusationCount
+            int accusationCount
     ) {
         return DisasterReportStatisticsResponse.builder()
                 .disasterReportId(disasterReportId)
